refactor: AuthHelper로 이름 변경 및 위치를 도메인 밖으로 이동하여 재사용
