+++ src/main/java/com/ddobang/backend/domain/AuthHelper.java
@@ -1,4 +1,4 @@
-package com.ddobang.backend.domain.party;
+package com.ddobang.backend.domain;
 
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
@@ -12,7 +12,7 @@ import lombok.RequiredArgsConstructor;
 // TODO: 임시 인증 헬퍼 클래스입니다. 멤버 도메인 쪽 인증 처리 완성되면 삭제 예정입니다
 @Component
 @RequiredArgsConstructor
-public class PartyAuthHelper {
+public class AuthHelper {
 
 	private final MemberRepository memberRepository;
