+++ src/main/java/com/ddobang/backend/domain/board/controller/AdminBoardController.java
@@ -1,0 +1,4 @@
+package com.ddobang.backend.domain.board.controller;
+
+public class AdminBoardController {
+}
