+++ src/main/java/com/ddobang/backend/domain/board/dto/response/PostDetailResponse.java
@@ -1,0 +1,4 @@
+package com.ddobang.backend.domain.board.dto.response;
+
+public class PostDetailResponse {
+}
