+++ src/main/java/com/ddobang/backend/domain/party/controller/PartyController.java
@@ -13,7 +13,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.ddobang.backend.domain.party.PartyAuthHelper;
+import com.ddobang.backend.domain.AuthHelper;
 import com.ddobang.backend.domain.party.dto.PartyDto;
 import com.ddobang.backend.domain.party.dto.request.PartyRequest;
 import com.ddobang.backend.domain.party.dto.request.PartySearchCondition;
@@ -35,7 +35,7 @@ import lombok.RequiredArgsConstructor;
 @Tag(name = "Party Controller")
 public class PartyController {
 	private final PartyService partyService;
-	private final PartyAuthHelper partyAuthHelper;
+	private final AuthHelper authHelper;
 
 	@GetMapping
 	@Operation(summary = "모임 목록 조회 (무한 스크롤)")
@@ -50,61 +50,61 @@ public class PartyController {
 	@GetMapping("/{id}")
 	@Operation(summary = "모임 상세 조회")
 	public ResponseEntity<SuccessResponse<PartyDetailResponse>> getParty(@PathVariable Long id) {
-		return ResponseFactory.ok(partyService.getPartyDetailResponse(id, partyAuthHelper.getCurrentMember()));
+		return ResponseFactory.ok(partyService.getPartyDetailResponse(id, authHelper.getCurrentMember()));
 	}
 
 	@PostMapping
 	@Operation(summary = "모임 등록")
 	public ResponseEntity<SuccessResponse<PartyDto>> createParty(@RequestBody @Valid PartyRequest request) {
-		return ResponseFactory.created(partyService.createParty(request, partyAuthHelper.getCurrentMember()));
+		return ResponseFactory.created(partyService.createParty(request, authHelper.getCurrentMember()));
 	}
 
 	@PutMapping("/{id}")
 	@Operation(summary = "모임 수정")
 	public ResponseEntity<SuccessResponse<PartyDto>> modifyParty(@PathVariable Long id,
 		@RequestBody @Valid PartyRequest request) {
-		return ResponseFactory.ok(partyService.modifyParty(id, request, partyAuthHelper.getCurrentMember()));
+		return ResponseFactory.ok(partyService.modifyParty(id, request, authHelper.getCurrentMember()));
 	}
 
 	@DeleteMapping("/{id}")
 	@Operation(summary = "소프트 딜리트 (모임 삭제)")
 	public ResponseEntity<Void> softDeleteParty(@PathVariable Long id) {
-		partyService.softDeleteParty(id, partyAuthHelper.getCurrentMember());
+		partyService.softDeleteParty(id, authHelper.getCurrentMember());
 		return ResponseFactory.noContent();
 	}
 
 	@PostMapping("/{id}/apply")
 	@Operation(summary = "모임 참가 신청")
 	public ResponseEntity<Void> applyParty(@PathVariable Long id) {
-		partyService.applyParty(id, partyAuthHelper.getCurrentMember());
+		partyService.applyParty(id, authHelper.getCurrentMember());
 		return ResponseFactory.noContent();
 	}
 
 	@DeleteMapping("/{id}/cancel")
 	@Operation(summary = "모임 참가 신청 취소")
 	public ResponseEntity<Void> cancelAppliedParty(@PathVariable Long id) {
-		partyService.cancelAppliedParty(id, partyAuthHelper.getCurrentMember());
+		partyService.cancelAppliedParty(id, authHelper.getCurrentMember());
 		return ResponseFactory.noContent();
 	}
 
 	@PostMapping("/{id}/accept/{memberId}")
 	@Operation(summary = "모임 신청 승인")
 	public ResponseEntity<Void> acceptPartyMember(@PathVariable Long id, @PathVariable Long memberId) {
-		partyService.acceptPartyMember(id, memberId, partyAuthHelper.getCurrentMember());
+		partyService.acceptPartyMember(id, memberId, authHelper.getCurrentMember());
 		return ResponseFactory.noContent();
 	}
 
 	@PatchMapping("/{id}/executed")
 	@Operation(summary = "모임 실행 완료")
 	public ResponseEntity<Void> executeParty(@PathVariable Long id) {
-		partyService.executeParty(id, partyAuthHelper.getCurrentMember());
+		partyService.executeParty(id, authHelper.getCurrentMember());
 		return ResponseFactory.noContent();
 	}
 
 	@PatchMapping("/{id}/unexecuted")
 	@Operation(summary = "모임 미실행 완료")
 	public ResponseEntity<Void> unexecuteParty(@PathVariable Long id) {
-		partyService.unexecuteParty(id, partyAuthHelper.getCurrentMember());
+		partyService.unexecuteParty(id, authHelper.getCurrentMember());
 		return ResponseFactory.noContent();
 	}
 }
