+++ src/test/java/com/ddobang/backend/domain/party/testUtils/MockConfig.java
@@ -1,6 +1,6 @@
 package com.ddobang.backend.domain.party.testUtils;
 
-import com.ddobang.backend.domain.party.PartyAuthHelper;
+import com.ddobang.backend.domain.AuthHelper;
 import org.mockito.Mockito;
 import org.springframework.boot.test.context.TestConfiguration;
 import org.springframework.context.annotation.Bean;
@@ -11,7 +11,7 @@ import org.springframework.context.annotation.Profile;
 @Profile("test")
 public class MockConfig {
     @Bean
-    public PartyAuthHelper partyAuthHelper() {
-        return Mockito.mock(PartyAuthHelper.class);
+    public AuthHelper authHelper() {
+        return Mockito.mock(AuthHelper.class);
     }
 }
